Organize refactor (#55)
* big ole refactor

* update Dockerfile

* remove unused constants
