+++ Dockerfile
@@ -8,9 +8,10 @@ ENV CGO_ENABLED=0 GOOS=linux GO111MODULE=on
 WORKDIR /go/src/github.com/virtual-vgo/vvgo
 COPY go.mod go.sum ./
 RUN go mod download
-COPY . .
-RUN go generate \
-    && go build -v -o /vvgo
+
+COPY . ./
+RUN go generate ./... \
+    && go build -v -o /vvgo github.com/virtual-vgo/vvgo/cmd/vvgo
 
 FROM builder as tester
 CMD ["go", "test", "-v", "./..."]
